worker: delete btrfs volume if it exists
but only when using the overlay driver.

We found that if you ended up in a state where you were trying to run a
worker with overlay and created the overlay mounts intside a btrfs
mount, lots of weird stuff would happen with volume creation and COW
volumes specifically.

This commit ensures that if you're using overlay and the volume path is
a btrfs mount, then btrfs mount is cleaned up.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>